Upgrade to Rails 5 and Bootstrap 4
Major overhaul of the styles, and minor changes to the rails app.

Wanted to move away from bitters and bourbon because their versions are
highly dependant on the sass gem, which made it harder to upgrade rails
without also having to update the styling. Because of this, and the fact
I didn't want to maintain a bunch of hand rolled CSS, I moved things t
Bootstrap 4.