Remove left and right padding from navbar
The navbar is nested in a container so that it's contents will have the
same width as the content below it. But the default bootstrap styles add
a padding to keep the contents indented from the sides of the window. I
don't need that since I have the content div as a parent. This also keeps
things lined up nicely.