Add Drag-n-drop file uploads to wiki textarea form elements
Inline Attachment
-----------------

Implements the Inline Attachment js library, so we can drag and drop
images directly into the wiki editing textarea (ala GitHub issues).

Originally I wanted to do the direct uploads, but in looking at
integrating that with Inline Attachment ruled that out. These changes
have the uploads proxied through the rails app directly. While they will
block a web worker during the upload this is much more flexible and
configurable.

In order to support Inline Attachment, I needed to add an `Upload`
resource, which will create a row in the uploads table and can process
the file to which ever service we want (currently S3 and Filesystem are
supported).

If the app is deployed to something like Heroku then you'll need S3 to
store the file. But in development (or on a server with a file system
that survives reboots) you can just use the filesystem.

S3 Uploads:

- Required Settings:
  - Access Key
  - Access Secret
  - Bucket Name
  - Region Name
- Receive the upload from the multi-part POST
- Generate a unique key for the file
- POST the file to S3, blocking the request
- Save the URL in the DB

Filesystem Uploads

- Receive the upload from the multi-part POST
- Generate a unique name for the file
- Move the temp file to `RAILS_ROOT/public/uploads`
- Save the URL in the DB

Upload Configuation and Initializer
-----------------------------------

Everything can be configured using enviroment variables. Basically you
can configure where you want the uploads to go, and provide any settings
needed to get them there. See `.env-example` for details.

Originally I wanted to use a class variable that would get set in the
initializer, however I think the Rails autoloading was breaking that. So
changed this to use Rails.configuration` setting instead. This seemed
appropprate enough, but a little clunky. My testing so far has been
stable with that setting.

Upload Resource
---------------

You can upload files either via the Inline Upload javascript, or via an
upload page. This change doesn't include a link to the upload page in
the web UI, but by visiting '/uploads/new' you can do it anyway. The
upload resource is not mutable, so you can create and delete them. No
editing at all.

Upload index is paginated with kaminari. Unfortunatly this gem didn't
work with Rails 5 on the released version so I had to use the latest
version on master from their GitHub repo. I considered removing that
from this change infavor of a low-tech hand-rolled version but this was
in and working. So YOLO?