Upload processor, page versions and profiles
This is a fairly sweeping change that touches quite a bit. Getting
things closer to being stable, but also just doing general cleanup as we
go.

If there is no upload processor set it to nil
---------------------------------------------

I had been using the null upload processor, but that seemed a bit
overengineered. So I ripped it out in favor of just setting the class to
nil. There is only one check for that -- and then uploads should be
disabled in the web UI.

If there are any leaks for where it is not disabled I would like to fix
that instead of making the upload and processor classes more
complicated.

Support for multiple versions of the wiki pages
-----------------------------------------------

This uses the paper trail gem to add change history to the wiki pages.

https://github.com/airblade/paper_trail

Paper Trail simply serializes the model at the time of save to it's own
table. This makes it very easy to integrate.

Also added a way to view old revisions and tweaked with some of the
layout and styling.

Add profiles and avatars
------------------------

Google oauth will return an image when logging the user in. That image
url is saved in the `avatar_url` column in the users table. Even display
that fancy pic in the navbar when the user is logged in.