Add page preview tab
The preview tab will make an AJAX post to the server that will render
the markdown and return html. While this is probably slower than
rendering things in JS, it should always product the same markdown
output as is used when rendering the page normally. Including fancy
things like emoji and tables.