Select files to upload
This allows uploading of files by clicking and browsing for files to
upload. This does some EJS-like processing:

On `pages/_form` we added:

  - hidden upload form
  - hidden iframe

Then when the user needs to upload a file it is posted to the iframe
which submits the original action (via the frame target attribute).

The form has a special strategy parameter to indicate to the action that
it is special, and should render a view that is expected in this iframe
post.

The upload is processed and the special view is rendered in the iframe.
This contains some javascript that is run as the page is loaded to call
back to the parent page and indicate the upload has processed.

That callback (parent.handle_iframe_upload) will then take the details
of the upload (in this case "details" means the markdown link) and
update the text area for the page.