Upgrade Rails, Ruby, and switch to Kramdown
The older Redcarpet and Pygments.rb libraries seem to be deprecated so
this removes them in favor of the Kramdown gem. Added bonus that this
seems to speed up and simplify the generation of HTML from the
markdown/source files.

Most of the changes for this were in the Renderer class and mucking with
style sheets to make those match up with the newly generated class
names. Luckily I could use the Rouge highlighter which can generate
Pygments-compatible css classes for the code-itself. So the highlight
color scheme continues to work. The built in default would inline the
styles into the html, and configuring was more work than switching
highlighters.

Also only include the `aws-sdk-s3` gem, since inclusing `aws-sdk` pulled
in dozens of gems.