Upgrade Rails to 6.1
As part of this we needed to do a couple of things:

Change the DB YAML format to not use references. Not sure why this
caused crashes, but it did and was quicker to remove those than debug
the YAML library.

Update Omniauth flow to use a POST when logging in, mostly needed to
update the welcome auth page to use a form instead of a link.

Converted coffeescript files to JS