Don't redirect to the auth page when logging in
This was a problem when logging out. The log out process would send
users to the login page and add that page to the redirect to origin. So
logging back in would be at the same unhelpful login page but with a
message saying "signed in"

Ain't pretty logic, but damn if it don't work.