Add Mermaid Support
This adds support for [Mermaid.js][1] so we can render graphs in the wiki
pages by entering the mermaid markup in fenced code blocks:

    ```mermaid
    graph TD
    A[Mardown Example] --> B[Fancy Graph]
    ````

This works because code blocks with the "mermaid" language will be
rendered into a `<div class=\"mermaid\">` tag, which the mermaid.js will
detect and replace the contents in the DOM with the SVG of the graph.

Since I couldn't add multiple syntax highlighters to Kramdown I had to
make a new one that would wrap the existing Rouge syntax highlighter,
intercept anything with Mermaid, and pass on the rest to Rouge. Kinda
clunky, but seems to work well.

Preview was a little tricky since the rendered markdown is dynamically
injected into the page. But turns out you can call `mermaid.init()`
multiple times and it will happily re-run on the page.

[1]: https://mermaid-js.github.io