Upgrade Rails to 7.0.3.1
Fixes https://github.com/csexton/corporate-tool/security/dependabot/43

Active Record RCE bug with Serialized Columns

When serialized columns that use YAML (the default) are deserialized,
Rails uses YAML.unsafe_load to convert the YAML data in to Ruby objects.
If an attacker can manipulate data in the database (via means like SQL
injection), then it may be possible for the attacker to escalate to an
RCE.

There are no feasible workarounds for this issue, but other coders (such
as JSON) are not impacted.