Upgrade rails to 7.0.4.1
Fixes https://github.com/csexton/corporate-tool/security/dependabot/58

iSQL Injection Vulnerability via ActiveRecord commentso

There is a possible vulnerability in ActiveRecord related to the
sanitization of comments. This vulnerability has been assigned the CVE
identifier CVE-2023-22794.

Versions Affected: >= 6.0.0 Not affected: < 6.0.0 Fixed Versions:
6.0.6.1, 6.1.7.1, 7.0.4.1

Impact

Previously the implementation of escaping for comments was insufficient
for

If malicious user input is passed to either the annotate query method,
the optimizer_hints query method, or through the QueryLogs interface
which automatically adds annotations, it may be sent to the database
with insufficient sanitization and be able to inject SQL outside of the
comment.